requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
sentence-transformers
langchain
langchain-community
langchain-huggingface
unstructured
unstructured[pdf]
faiss-cpu
python-magic
python-magic-bin
requests
httpx